Split page into components, hooks etc.
