+++ components/systems/Card.tsx
@@ -1,0 +1,3 @@
+const Card = (props: any) => <div {...props} className={`mb-2 lg:mb-4 py-1 lg:py-2 ${props.className}`} />
+
+export default Card
